+++ boop/component/component.py
@@ -1,11 +1,13 @@
 from .componenthost import ComponentHost
 
-
 class Component(ComponentHost):
     def __init__(self):
         ComponentHost.__init__(self)
+        self.state_stack = list()
 
     def handle_event(self, event_type, state, *args, **kwargs):
+        # FIXME we should handle exceptions to an extent here
+        self.state_stack.append(state)
         evthandler = None
         try:
             evthandler = self.__getattribute__('do_'+event_type)
@@ -18,4 +20,6 @@ class Component(ComponentHost):
         if evthandler:
             evthandler(state, *args, **kwargs)
 
-        return ComponentHost.handle_event(self, event_type, state, *args)
+        ret = ComponentHost.handle_event(self, event_type, state, *args)
+        self.state_stack.pop()
+        return ret